[repo] docs: Added a link to "How to create and deploy models and embeddings" (#2453)
## Issue 
Fixes #2465 

## Details

Just added a link to community article that walks the developer through
the steps of creating and deploying embeddings and models in Azure
portal.

#### Change details

Just added a link to community article that walks the developer through
the steps of creating and deploying embeddings and models in Azure
portal.

Link that I added:
https://techcommunity.microsoft.com/blog/startupsatmicrosoftblog/how-to-set-up-and-configure-a-gpt-deployment-using-the-azure-openai-service/3849854

## Attestation Checklist

- [x] My changes follows the style guidelines of this project
- [x] I have checked for/fixed spelling, linting, and other errors
- [x] My changes generate no new warnings

---------

Co-authored-by: Gaurav Keshre <gakeshre@microsoft.com>