Add scripts to generate llmstxt files (#2513)
LLMs txt (https://llmstxt.org/#format) is a format that's being adopted
in the industry to help coding assistants leverage the library
effectively.

This PR introduces a script to generate llms.txt (small) and llms.txt
(full) for both typescript and C#. In addition it generates a txt file
for each document.
llms.txt (small) - this is essentially an index file that contains links
to other txt files
llms.txt (full) - this is essentially a full dump of the entire
documentation all concatenated together.

The files get generated when the deploy is happening. 
Also added an actual documentation to help folks discover these files.

The scripts are basically vibecoded, but I've gone through it and they
seem to be correct, and doing what I want.