Remove submodules in favor of linking to README (#2651)
Submodules are a bit of a cumbersome to keep up to date. It requires us
to manually run `git submodule update --init --recursive --remote"` to
keep those links up to date. And they don't necessarily provide any
super useful value.
So in this PR we are removing them and simply moving them as links in
the README.